requirements.txt
pandas~=1.4.2
numpy~=1.22.4
lxml~=4.9.0
pyunpack~=0.3
selenium~=4.2.0
beautifulsoup4~=4.11.1
biopython~=1.79
html5lib~=1.1
webdriver_manager~=3.7.0
requests~=2.28.0
PySide6~=6.3.1
coloredlogs~=15.0.1
PyYAML~=6.0
biopython~=1.79